[BE/ENV] 음식 정보 DB에 입력

+++ backend/fns/src/main/java/ssafy/fns/domain/exercise/service/dto/ExerciseResponseDto.java
@@ -14,5 +14,5 @@ public class ExerciseResponseDto {
 
     private String sportsName;
 
-    private Long kcal;
+    private Double kcal;
 }
